docs(dsh): one example end to end — ram + redis + slack, monty capturing python
